requirements.txt
Flask==3.0.0
Flask-CORS>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
soundfile>=0.12.0
Pillow>=10.0.0
numpy>=1.26.0
pandas>=2.1.3
scikit-learn>=1.3.0
matplotlib>=3.7.0
requests>=2.31.0
reportlab>=4.0.0
